Refactor import paths: Update all imports from 'lspo_3d' to 'src.lspo_3d' for consistency
