+++ stls/src/lspo_3d/models/generator.py
@@ -13,7 +13,7 @@ import torch
 import torch.nn as nn
 from torch import Tensor
 
-from lspo_3d import config
+from src.lspo_3d import config
 
 
 class CSGGenerator(nn.Module):
