+++ stls/src/lspo_3d/rl/agent.py
@@ -21,7 +21,7 @@ from stable_baselines3.common.policies import ActorCriticPolicy
 from stable_baselines3.common.callbacks import BaseCallback
 
 if typing.TYPE_CHECKING:
-    from lspo_3d import config
+    from src.lspo_3d import config
     from lspo_3d.rl.environment import DesignEnvironment
